+++ src/app/plans/plans/plans.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { StripeService } from 'src/app/services/stripe.service';
+import { loadStripe } from '@stripe/stripe-js';
+import { environment } from '../../../environments/environment';
 
 @Component({
   selector: 'app-plans',
@@ -7,11 +9,25 @@ import { StripeService } from 'src/app/services/stripe.service';
   styleUrls: ['./plans.component.scss'],
 })
 export class PlansComponent implements OnInit {
+  sessionId;
   constructor(private stripeService: StripeService) {}
 
-  ngOnInit(): void {}
+  ngOnInit(): void {
+    this.stripeService.addCheckoutSession().then((ref) => {
+      ref.onSnapshot((snap) => {
+        const { error, sessionId } = snap.data();
+        if (error) {
+          alert(`エラーが発生しました: ${error.message}`);
+        }
+        if (sessionId) {
+          this.sessionId = sessionId;
+        }
+      });
+    });
+  }
 
-  redirectToCheckout() {
-    this.stripeService.redirectToCheckout();
+  async redirectToCheckout() {
+    const stripe = await loadStripe(environment.stripe.publicKey);
+    stripe.redirectToCheckout({ sessionId: this.sessionId });
   }
 }
